refactor(logg): rename repo to dao
- Rename repo package to dao in services/logg
- Update import paths and variable types from repo.ILogg to dao.ILogg
- Modify UseDb function to accept dao.ILogg instead of repo.ILogg
