+++ services/logg/dao/ilogg.go
@@ -1,4 +1,4 @@
-package repo
+package dao
 
 import (
 	"context"
