+++ services/logg/dao/impl.go
@@ -1,4 +1,4 @@
-package repo
+package dao
 
 import (
 	"context"
