+++ services/logg/dao/service.go
@@ -1,4 +1,4 @@
-package repo
+package dao
 
 import (
 	"fmt"
